将ojrunner改造为完全在docker内执行，外部通过http访问

+++ site/ojrunner/shell/compile.sh
@@ -33,6 +33,10 @@ case "$compiler" in
 		"$compiler" "$srcPath" -o "$buildPath";;
 esac
 
+if [ "$?" -ne 0 ];then
+	exit 1
+fi
+
 #清理中间文件（.o文件）
 if [ -f "${buildPath}.o" ];then
 	rm "${buildPath}.o"
